Add automatic picolibc build
Add a script to build picolibc by using make picolibc. The script works with riscv32 and riscv64 compilers.
It is installed locally to the repository with a specs-specs file containing all necessary options for correct linking.
Also a dummyfile is included for printf support which needs to be adjusted in future commits.